move v2components
