+++ pages/v2/_skeletons/aquired.tsx
