+++ pages/v2/_skeletons/aquired/[uniqueKey].tsx
