+++ pages/v2/_skeletons/boosters.tsx
