+++ pages/v2/_skeletons/rewards.tsx
