+++ pages/v2/_skeletons/rewards/[uniqueKey].tsx
