+++ pages/v2/_skeletons/user-registration.tsx
