+++ pages/v2/_skeletons/waitings.tsx
